create two openocd-backed ITM examples
japaric/itm-tools has now been verified to work on the stm32f401re. Next
up we want to move over from openocd to probe-rs. Before we start
investigating it, however, we shall port the openocd commands

    monitor mmw 0xE0001000 65536 4096
    monitor itm port 0 on

to Rust code that we can put in init(). This will allow us to (almost)
drop-in replace openocd with probe-rs once we move over. What we may
struggle with in probe-rs is the ITM packet receiving; we ultimately
want to replicate

    monitor tpiu config internal itm.bin uart off 16000000

When the above has been done, we can fork japaric/itm-tools and provide
probe-rs instructions instead. This also forces me to find ARM document
references for all relevant registers and concepts, in case `monitor
tpiu ...` functionality is already available in probe-rs.