rtic-trace/macros: inline set_current_task_id
We want the least amount of overhead when tracing the user application.
So eventually we'll have to asm-inspect all code that is injected into
it, set_current_task_id being one of them. Obviously, we don't want to
pay for function entries and exits, so we suggest that the function
should be inlined.