rtic-trace: improve hardware trace function name
In the context of the crate in a void, it is not clear what a task is
nor what its ID would denote. In practise we write a payload to a
register and expect side-effects. The name change reflect this.
Additionally, a __ prefix has been added to the function name to denote
its internal usage. Unfortunately, we cannot hide it from the user.