refactor: merge rtic_trace::parsing into cargo-rtic-trace
rtic_trace::parsing is a non-trivial feature-set currently only in use
by cargo-rtic-trace. To mix it with a no-std feature-set and make the
two sets mutually exclusive is just asking for trouble. In the end we
want to minimize the overhead of the end-user. If a crate must then be
included it should not require any features.

If needed, the parsing module can later be split out into some
cortex-m-rtic-trace-syntax crate.